Day 21
Part1 - Pretty easy, able to get it done no problem, only keeps the current state of where you are and copies a fresh map back to then receive the new positions

Part2 - Not fun, took some time, did see the empty spaces that run up and down and along the input. Figured there was a way to do it with math, since running through all the steps and especially with infinite size would of been a pain. Saw some tips that using the quadratic fomula would be useful